// src/api/models/leader.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend
 * Coded by Meu TEAM
 * OpenAPI spec version: 1.0.0
 */
import type { LeaderAvatarSummary } from './leaderAvatarSummary';

export interface Leader {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  position: string | null;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  avatar_id: string | null;
  /** @nullable */
  avatar: LeaderAvatarSummary;
  /** @nullable */
  linkedin_url: string | null;
  /** @nullable */
  started_at: string | null;
  /** @nullable */
  ended_at: string | null;
  display_order: number;
  is_active: boolean;
  created_at: string;
  /** @nullable */
  created_by?: string | null;
  updated_at: string;
  /** @nullable */
  updated_by?: string | null;
}